kafka_spark_stream/requirements.txt
astroid==2.4.2
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
colorama==0.4.4
idna==2.10
isort==4.3.21
joblib==0.17.0
kafka-python==1.4.7
lazy-object-proxy==1.4.3
mccabe==0.6.1
nltk==3.5
oauthlib==3.1.0
psutil==5.7.3
py4j==0.10.7
pylint==2.5.3
PySocks==1.7.1
pyspark==2.4.4
regex==2020.11.13
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
textblob==0.15.3
toml==0.10.1
tqdm==4.54.0
tweepy==3.8.0
typed-ast==1.4.1
urllib3==1.26.2
wrapt==1.12.1
